agent_installer/requirements_agent.txt
sounddevice
webrtcvad-wheels
requests
python-dotenv
numpy
